pyproject.toml
[tool.poetry]
name = "p5"
version = "0.1.0"
description = "Práctica 5 de Gestión de Datos (2023-24)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
kafka-python = "^2.0.2"
matplotlib = "^3.8.4"

[tool.poetry.scripts]
start = 'p5.p5:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
